% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_v_characteristic.R
\name{iv.plot}
\alias{iv.plot}
\title{I-V curve Function}
\usage{
iv.plot(dane)
}
\arguments{
\item{dane}{data table with `voltage` and `current_na_cm2`}
}
\value{
plot of i-v characteristic
}
\description{
This funcion draws the current-voltage charasteristic
}
\keyword{curve}
\keyword{i-v}
