% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_v_characteristic.R
\name{iv.summary}
\alias{iv.summary}
\title{Summary of solar cells characteristics}
\usage{
iv.summary(data_dir, output_dir, pattern = "")
}
\arguments{
\item{data_dir}{input directory}

\item{output_dir}{output directory}

\item{pattern}{which files to select}
}
\value{
data frame containing i-v summary (Isc, Voc, FF, efficiency, resistivity)
}
\description{
Given an input directory `data_dir` the function looks for `.dat` files matching `patern`
and uses them to calculate and plot i-v characterisitcs
}
\keyword{cell}
\keyword{curve,}
\keyword{i-v}
\keyword{parameters}
\keyword{solar}
