% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_v_characteristic.R
\name{PVparameters}
\alias{PVparameters}
\title{Solar cell parameters}
\usage{
PVparameters(dane)
}
\arguments{
\item{dane}{data table with `voltage` and `current_na_cm2`}
}
\value{
vector containing i-v summary (Isc, Voc, FF, efficiency, resistivity)
}
\description{
For a given current - voltage data this function calculates the maximal power (Pmax) using interpolation,
efficiency, short current denisty (I_sc), open circuit voltage (V_oc), FF factor, resistivity
}
\keyword{cell}
\keyword{parameters}
\keyword{solar}
